84.deploy-using-python/deploy-using-python/requirements.txt
paramiko==3.5.0
python-decouple==3.8
